Fixes for hashes and exponentiation to better support 32-bit builds

+++ src/Symbolics.jl
@@ -733,6 +733,6 @@ include("despecialize.jl")
     end
 end
 
-precompile(Tuple{typeof(Base.:(^)), SymbolicUtils.BasicSymbolicImpl.var"typeof(BasicSymbolicImpl)"{SymbolicUtils.SymReal}, Int64})
+precompile(Tuple{typeof(Base.:(^)), SymbolicUtils.BasicSymbolicImpl.var"typeof(BasicSymbolicImpl)"{SymbolicUtils.SymReal}, Int})
 
 end # module
